refactor: service/serviceImpl로 구조 변경
