+++ src/main/java/project/trendpick_pro/domain/category/service/SubCategoryService.java
@@ -14,52 +14,13 @@ import java.util.List;
 
 ;
 
-@Service
-@RequiredArgsConstructor
-@Transactional(readOnly = true)
-public class SubCategoryService {
 
-    private final SubCategoryRepository subCategoryRepository;
-
-    @Transactional
-    public void save(String name, MainCategory mainCategory){
-        subCategoryRepository.save(new SubCategory(name, mainCategory));
-    }
-
-    @Transactional
-    public void saveAll(List<String> name, MainCategory mainCategory){
-        List<SubCategory> list = new ArrayList<>();
-        for (String s : name) {
-            list.add(new SubCategory(s, mainCategory));
-        }
-        subCategoryRepository.saveAll(list);
-    }
-
-    @Transactional
-    public void delete(Long id){
-        SubCategory subCategory = subCategoryRepository.findById(id).orElseThrow();
-        subCategoryRepository.delete(subCategory);
-    }
-    public List<SubCategoryResponse> findAll(String mainCategoryName) {
-        if (mainCategoryName.equals("전체")){
-            List<SubCategory> categories = subCategoryRepository.findAllBy();
-            return categories.stream().map(subCategory -> new SubCategoryResponse(subCategory.getName())).toList();
-        } else {
-            List<SubCategory> categories = subCategoryRepository.findAllByMainCategory(mainCategoryName);
-            return categories.stream().map(subCategory -> new SubCategoryResponse(subCategory.getName())).toList();
-        }
-    }
-
-    public SubCategoryResponse findById(Long id){
-        SubCategory subCategory = subCategoryRepository.findById(id).orElseThrow();
-        return new SubCategoryResponse(subCategory.getName());
-    }
-
-    public SubCategory findByBaseId(Long id, MainCategory mainCategory){
-        return subCategoryRepository.findByIdInMainCategory(id, mainCategory);
-    }
-
-    public SubCategory findByName(String username){
-        return subCategoryRepository.findByName(username);
-    }
+public interface SubCategoryService {
+    void save(String name, MainCategory mainCategory);
+    void saveAll(List<String> name, MainCategory mainCategory);
+    void delete(Long id);
+    List<SubCategoryResponse> findAll(String mainCategoryName);
+    SubCategoryResponse findById(Long id);
+    SubCategory findByBaseId(Long id, MainCategory mainCategory);
+    SubCategory findByName(String username);
 }
